+++ cruncher/udp_flow_stats.go
@@ -29,7 +29,7 @@ type UDPFlowStats struct {
 			Blksize    int    `json:"blksize"`
 			Omit       int    `json:"omit"`
 			Duration   int    `json:"duration"`
-			Bytes      int    `json:"bytes"`
+			Bytes      int64  `json:"bytes"`
 			Blocks     int    `json:"blocks"`
 			Reverse    int    `json:"reverse"`
 			Tos        int    `json:"tos"`
@@ -41,7 +41,7 @@ type UDPFlowStats struct {
 			Start         float64 `json:"start"`
 			End           float64 `json:"end"`
 			Seconds       float64 `json:"seconds"`
-			Bytes         int     `json:"bytes"`
+			Bytes         int64   `json:"bytes"`
 			BitsPerSecond float64 `json:"bits_per_second"`
 			JitterMs      float64 `json:"jitter_ms"`
 			LostPackets   int     `json:"lost_packets"`
@@ -53,7 +53,7 @@ type UDPFlowStats struct {
 			Start         float64 `json:"start"`
 			End           float64 `json:"end"`
 			Seconds       float64 `json:"seconds"`
-			Bytes         int     `json:"bytes"`
+			Bytes         int64   `json:"bytes"`
 			BitsPerSecond float64 `json:"bits_per_second"`
 			JitterMs      float64 `json:"jitter_ms"`
 			LostPackets   int     `json:"lost_packets"`
@@ -69,7 +69,7 @@ type UDPFlowStats struct {
 				Start         float64 `json:"start"`
 				End           float64 `json:"end"`
 				Seconds       float64 `json:"seconds"`
-				Bytes         int     `json:"bytes"`
+				Bytes         int64   `json:"bytes"`
 				BitsPerSecond float64 `json:"bits_per_second"`
 				JitterMs      float64 `json:"jitter_ms"`
 				LostPackets   int     `json:"lost_packets"`
@@ -82,7 +82,7 @@ type UDPFlowStats struct {
 			Start         float64 `json:"start"`
 			End           float64 `json:"end"`
 			Seconds       float64 `json:"seconds"`
-			Bytes         int     `json:"bytes"`
+			Bytes         int64   `json:"bytes"`
 			BitsPerSecond float64 `json:"bits_per_second"`
 			JitterMs      float64 `json:"jitter_ms"`
 			LostPackets   int     `json:"lost_packets"`
@@ -128,7 +128,7 @@ type UDPFlowStats struct {
 				Blksize    int    `json:"blksize"`
 				Omit       int    `json:"omit"`
 				Duration   int    `json:"duration"`
-				Bytes      int    `json:"bytes"`
+				Bytes      int64  `json:"bytes"`
 				Blocks     int    `json:"blocks"`
 				Reverse    int    `json:"reverse"`
 				Tos        int    `json:"tos"`
@@ -140,7 +140,7 @@ type UDPFlowStats struct {
 				Start         float64 `json:"start"`
 				End           float64 `json:"end"`
 				Seconds       float64 `json:"seconds"`
-				Bytes         int     `json:"bytes"`
+				Bytes         int64   `json:"bytes"`
 				BitsPerSecond float64 `json:"bits_per_second"`
 				Packets       int     `json:"packets"`
 				Omitted       bool    `json:"omitted"`
@@ -149,7 +149,7 @@ type UDPFlowStats struct {
 				Start         float64 `json:"start"`
 				End           float64 `json:"end"`
 				Seconds       float64 `json:"seconds"`
-				Bytes         int     `json:"bytes"`
+				Bytes         int64   `json:"bytes"`
 				BitsPerSecond float64 `json:"bits_per_second"`
 				Packets       int     `json:"packets"`
 				Omitted       bool    `json:"omitted"`
@@ -162,7 +162,7 @@ type UDPFlowStats struct {
 					Start         float64 `json:"start"`
 					End           float64 `json:"end"`
 					Seconds       float64 `json:"seconds"`
-					Bytes         int     `json:"bytes"`
+					Bytes         int64   `json:"bytes"`
 					BitsPerSecond float64 `json:"bits_per_second"`
 					JitterMs      float64 `json:"jitter_ms"`
 					LostPackets   int     `json:"lost_packets"`
@@ -175,7 +175,7 @@ type UDPFlowStats struct {
 				Start         float64 `json:"start"`
 				End           float64 `json:"end"`
 				Seconds       float64 `json:"seconds"`
-				Bytes         int     `json:"bytes"`
+				Bytes         int64   `json:"bytes"`
 				BitsPerSecond float64 `json:"bits_per_second"`
 				JitterMs      float64 `json:"jitter_ms"`
 				LostPackets   int     `json:"lost_packets"`
